Refactor scmBuild commands out of siteTreeitem
