+++ src/commands/deploy/deploy.ts
@@ -18,6 +18,8 @@ import { isPathEqual } from '../../utils/pathUtils';
 import { getRandomHexString } from "../../utils/randomUtils";
 import { getWorkspaceSetting } from '../../vsCodeConfig/settings';
 import { runPostDeployTask } from '../postDeploy/runPostDeployTask';
+import { promptScmDoBuildDeploy } from '../promptScmDoBuildDeploy';
+import { promptToSaveDeployDefaults } from '../promptToSaveDeployDefaults';
 import { confirmDeploymentPrompt } from './confirmDeploymentPrompt';
 import { getDeployNode, IDeployNode } from './getDeployNode';
 import { IDeployContext, WebAppSource } from './IDeployContext';
@@ -69,9 +71,9 @@ export async function deploy(context: IActionContext, target?: vscode.Uri | Site
             const linuxFxVersion: string = siteConfig.linuxFxVersion.toLowerCase();
             if (linuxFxVersion.startsWith(appservice.LinuxRuntimes.node)) {
                 // if it is node or python, prompt the user (as we can break them)
-                await node.promptScmDoBuildDeploy(deployContext.effectiveDeployFsPath, appservice.LinuxRuntimes.node, context);
+                await promptScmDoBuildDeploy(deployContext, deployContext.effectiveDeployFsPath, appservice.LinuxRuntimes.node);
             } else if (linuxFxVersion.startsWith(appservice.LinuxRuntimes.python)) {
-                await node.promptScmDoBuildDeploy(deployContext.effectiveDeployFsPath, appservice.LinuxRuntimes.python, context);
+                await promptScmDoBuildDeploy(deployContext, deployContext.effectiveDeployFsPath, appservice.LinuxRuntimes.python);
             }
 
         }
@@ -82,7 +84,7 @@ export async function deploy(context: IActionContext, target?: vscode.Uri | Site
     }
 
     // tslint:disable-next-line:no-floating-promises
-    node.promptToSaveDeployDefaults(deployContext, deployContext.workspace.uri.fsPath, deployContext.effectiveDeployFsPath);
+    promptToSaveDeployDefaults(deployContext, node, deployContext.workspace.uri.fsPath, deployContext.effectiveDeployFsPath);
     await appservice.runPreDeployTask(deployContext, deployContext.originalDeployFsPath, siteConfig.scmType);
 
     // cancellation moved to after prompts while gathering telemetry
